@@ -1,10 +1,10 @@
 module github.com/odysseia-greek/delphi/ptolemaios
 
-go 1.20
+go 1.21
 
 require (
-	github.com/odysseia-greek/agora/diogenes v0.1.0
-	github.com/odysseia-greek/agora/plato v0.1.1
+	github.com/odysseia-greek/agora/diogenes v0.1.4
+	github.com/odysseia-greek/agora/plato v0.1.4
 	github.com/stretchr/testify v1.8.4
 	google.golang.org/grpc v1.59.0
 	google.golang.org/protobuf v1.31.0
